military-law-assistant/requirements.txt
# Core Flask & Web Framework
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
blinker==1.8.2
itsdangerous==2.2.0
click==8.1.7
Flask-Cors==4.0.1

# AI & Machine Learning
openai>=1.12.0
pydantic==2.7.1
pydantic_core==2.18.2
annotated-types==0.6.0
typing-extensions>=4.11.0
tqdm==4.66.2

# Audio Processing
SpeechRecognition==3.10.1
pydub==0.25.1
pypdfium2==4.30.0
numpy==1.26.4
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# PDF Processing
pdfplumber==0.11.0
pdfminer.six==20231228
Pillow==10.3.0
PyPDF2==3.0.1

# HTTP & Networking
requests==2.31.0
httpx==0.27.0
httpcore==1.0.5
h11==0.14.0
urllib3==2.2.1

# Security
cryptography==42.0.5
certifi==2024.2.2

# Utilities
python-dotenv==1.0.0
distro==1.9.0
colorama==0.4.6
charset-normalizer==3.3.2
idna==3.6
MarkupSafe==2.1.5
sniffio==1.3.0
anyio==4.3.0
loguru==0.7.2
python-json-logger==2.0.7

# Production Server
waitress==3.0.0
gunicorn==21.2.0

# Monitoring
sentry-sdk==1.40.6
prometheus-client==0.20.0
flask-healthz==1.0.1  # Fixed version

# PWA Support (Optional)
Flask-PWA==1.2.0
manifest-json==1.1.0